feature(frontend): add BudgetList to index.vue
