+++ frontend/types/budget-list-response.d.ts
@@ -1,0 +1,7 @@
+import { Budget } from './budget'
+
+export interface BudgetListResponse {
+  next: string
+  prev: string
+  results: Budget[]
+}
