+++ frontend/types/budget.d.ts
@@ -1,0 +1,11 @@
+export interface Budget {
+  id: string
+  name: string
+  slug: string
+  year: number
+  subtitle: string
+  classificationSystem: string
+  government: string
+  createdAt: string
+  updatedAt: string
+}
